register, approve-registration, login (base codes) 01
OUTPUTS:
01:
I registered and it went to RegistrationRequest DB
I approved it using approve-registation ts and tsx
After approval it went to Residents and User DB
It also generated digital ID

	WHAT I WANT TO ADD NOW:
Uploading of photo during registration
Add “Are you head of the family?” in registration form if not then they must type the head ID of the head of the family (this should help for the household ID generation)
Things to add to form: gender, address, and ask if member of 4ps (since birthday is already there then the system must automatically determine if the resident is senior citizen or not) these should help for DemographicTags
Digital ID has the QR code and photo, when QR Code is scanned the resident’s information must be displayed.
Once request is approved it should be automatically removed in RegistrationRequest DB.
Must send an email/text once registration is approved. A message informing them their account is ready and their initial username and password is given.